Package theta_sweep module and add project metadata
